# openapi/music-intelligence-api.yaml
openapi: 3.0.3
info:
  title: Music Intelligence API
  version: 1.0.0
  description: |
    The Music Intelligence API provides real-time and predictive analytics for
    tracks and artists across Spotify, TikTok, YouTube, Apple Music, and radio.

    All endpoints require an API key passed via the `x-api-key` header.
    Contact your account manager to obtain credentials and configure scopes.

    ## Authentication
    Pass your API key in every request:
    ```
    x-api-key: YOUR_API_KEY
    ```

    ## Rate Limiting
    Default rate limit is 1,000 requests per minute per tenant.
    Exceed this and you will receive HTTP 429.

    ## Response Envelope
    All successful responses are wrapped:
    ```json
    { "obj": <data>, "offset": 0, "total": 100 }
    ```
    Errors:
    ```json
    { "error": "message" }
    ```

    ## BigInt Fields
    Large counters (streams, views, followers) are returned as **strings**
    to avoid JavaScript precision loss.
servers:
  - url: https://api.music-intelligence.io
    description: Production
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/tracks/{id}:
    get:
      summary: Get track metadata and stats
      description: Returns metadata, platform statistics, and external platform IDs for a single track.
      tags: [Tracks]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/TrackId'
      responses:
        '200':
          description: Track found
          content:
            application/json:
              schema:
                type: object
                properties:
                  obj:
                    $ref: '#/components/schemas/TrackResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/tracks/{id}/trend:
    get:
      summary: Get ML trend prediction for a track
      description: |
        Returns the latest ML-generated trend classification for a track.
        The model classifies tracks as `VIRAL`, `TRENDING`, `POPULAR`, or `NONE`
        using audio features, playlist signals, and UGC velocity from TikTok and Spotify.
      tags: [Tracks]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/TrackId'
      responses:
        '200':
          description: Trend prediction found
          content:
            application/json:
              schema:
                type: object
                properties:
                  obj:
                    $ref: '#/components/schemas/TrackTrendResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/artists/breaking:
    get:
      summary: List breaking and emerging artists
      description: |
        Returns artists ranked by breakout probability. Filter by status,
        genre, or country. Useful for talent discovery and A&R workflows.
      tags: [Artists]
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [ABOUT_TO_BREAK, GROWING, STABLE, DECLINING]
          description: Filter by trajectory status.
        - in: query
          name: genre
          schema:
            type: string
          description: Filter by primary genre (e.g. `hip-hop`, `pop`).
        - in: query
          name: code2
          schema:
            type: string
          description: Filter by ISO alpha-2 country code (e.g. `US`, `GB`).
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Artists returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  obj:
                    type: array
                    items:
                      $ref: '#/components/schemas/BreakingArtistRow'
                  offset:
                    type: integer
                  total:
                    type: integer
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  parameters:
    TrackId:
      in: path
      name: id
      required: true
      schema:
        type: integer
      description: Internal track ID.
...
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    TrackResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        isrc:
          type: string
          nullable: true
        externalIds:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Platform IDs keyed by platform name (e.g. spotify, apple_music, tiktok).
        artists:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
        statistics:
          type: object
          properties:
            raw:
              type: object
              properties:
                spotifyStreams:
                  type: string
                  nullable: true
                  description: Spotify streams (string to preserve BigInt precision).
                spotifyPopularity:
                  type: integer
                  nullable: true
                tiktokVideoCount:
                  type: integer
                  nullable: true
                youtubeViews:
                  type: string
                  nullable: true
            derived:
              type: object
              properties:
                streamVelocity7d:
                  type: number
                  nullable: true
                viralScore:
                  type: number
                  nullable: true
    TrackTrendResponse:
      type: object
      properties:
        trackId:
          type: integer
        prediction:
          type: object
          nullable: true
          properties:
            label:
              type: string
              enum: [VIRAL, TRENDING, POPULAR, NONE]
            probViral:
              type: number
              format: float
            probTrending:
              type: number
              format: float
            probPopular:
              type: number
              format: float
            probNone:
              type: number
              format: float
            modelName:
              type: string
            modelVersion:
              type: string
            createdAt:
              type: string
              format: date-time
        label:
          type: object
          nullable: true
          description: ETL-generated ground-truth label (used for model training).
          properties:
            label:
              type: string
              enum: [VIRAL, TRENDING, POPULAR, NONE]
            date:
              type: string
              format: date
...
    BreakingArtistRow:
      type: object
      properties:
        artistId:
          type: string
        name:
          type: string
          nullable: true
        code2:
          type: string
          nullable: true
        primaryGenre:
          type: string
          nullable: true
        primaryCode2:
          type: string
          nullable: true
        status:
          type: string
          enum: [ABOUT_TO_BREAK, GROWING, STABLE, DECLINING]
        statusScore:
          type: number
        breakProbability:
          type: number
        streams28dDelta:
          type: number
        playlistsDelta28d:
          type: number
          nullable: true
        followersDelta28d:
          type: number
          nullable: true
        tiktokVelocityScore:
          type: number
          nullable: true
        airplayVelocityScore:
          type: number
          nullable: true
        date:
          type: string
          format: date
...